feat: define custom promisify implementations
This defines properties on each of the exported asynchronous functions
which allow them to work with Node's `promisify`.

The existing callbacks were incompatible by default, because they:
  • are not given as the last arguments of those functions
  • do not accept optional "error" as their first arguments

In other words, they are not "Node-style".

This implementation chooses to reject promises rather than returning
`undefined` on those errors, as the callbacks do: presumably the
`undefined` value is used simply because the existing callback format
forgot to allow for an error parameter.

This defines both the modern `promisify.custom` symbol and the "old"
`__promisify__` property, since the former isn't actually describable in
TypeScript definitions yet, and the latter is still used in Node's
definition files (e.g. for `child_process` functions). See:
  • https://github.com/microsoft/TypeScript/issues/36813
  • https://github.com/DefinitelyTyped/DefinitelyTyped/pull/42154