nit: lift filter depth into filter arguments, not callback
This is a no-op, since the parameter was never given differently than
its default value, other than in recursive contexts.

That parameter was counter-abstract and only coincidentally common
between buildProcessTree and filterProcessList. If the value were ever
to be changeable, it should be as a bound parameter on the supplied
transform, e.g. `(pid, list) => buildProcessTree(pid, list, 320)`

Didn't belong on getRawProcessList.