fix: avoid orphaning chained callbacks
This fixes one of two causes of orphaned callbacks.

(The other cause is that only one of the two queues gets flushed after
each invocation).

Previously, if a callback might have called `getProcessList` or
`getProcessTree` again (which does *not* necessarily mean a recursion
problem), that callback would be added to the queue, but not processed,
and no `native.getProcessList` would be invoked. The callback would
remain orphaned until another invocation caused that queue to be flushed
again.

To avoid this, we repeat the draining until the queue is empty. We use
"queue.splice(0)" to atomically clear the queue, returning a batch of
callbacks to process. If any of those also made requests, we repeat
until the callback chain completes.

An alternative would be to splice the queue once and immediately reset
`requestInProgress` before invoking callbacks:
`CreateToolhelp32Snapshot` has safely completed at this point. However,
that would circumvent the "too many requests" rate-limiting (?) concern
mentioned in one of the comments.