perf: reduce redundant list searching
Previously `buildProcessTree` and `filterProcessList` performed several
redundant linear searches for each node in the process tree. At each
level:
 • a linear search to find the process matching the given pid
 • another linear search collecting a temporary list of children
 • then for each of those children (recursively), repeating the same
   searches over the entire list of processes.

This may not be performance-critical code, but it was needlessly
inefficient.

This updates the implementations to iterate over the list of processes
once, collecting the top-level "root" process and constructing a map of
from pid ⇒ children. Those lists of children are updated during
iteration and the tree is constructed by reference to those lists.