feat: export promisified variants under 'windows-process-tree/promises'
This adds the module 'windows-process-tree/promises', exposing the
promisified variants of the primary library methods.

It would be nice to also make this a property on the
'windows-process-tree' index module, but with the current structure that
requires some work to avoid a mutually-recursive import.

Since the `index.ts` file sets the [promisify.custom]/__promisify__
properties at the top level, and the './promises' module uses those at
the top level to expose the promisified implementations, there's an
initialization mutual dependency.

The simplest thing, done here, is to only support explicit importing of
`windows-process-tree/promises`.

To do it properly, either:
 • one of `index.ts` or `promises.ts` exposes lazy-init'd properties
   via a `defineProperty` getter on `exports`
 • or, simplest but noisiest, move the current `index.ts` to
   something like `_core.ts`, and create an `index.ts` with:
     export * from './_core';
     export * as promises from './promises';

However that doesn't don't seem worth immediate testing effort, and can
always be done later without breaking API compatibility.